Initial work on a special QP solver for SNS-IK
This code compiles but is entirely untested. It almost certainly
has some bug in it and is a work in progress.

This commit adds code that solves a quadratic program with a
diagonal quadratic cost, zero linear cost, equality constraints,
and bounds on the inputs. It uses a sparse linear solver to be
as fast as possible. The solver implements a simple active set
method for handling the limits on the decision variables.

The motivation here is that the SNS-IK paper outlines a solver
that uses a modified version of an active set method, but it is
difficult to understand and debug. It also is slightly different
for each version of the algorithm. I'm curious to see if a more-
general solver like this one can be used at the core of SNS-IK.