Currently failing unit tests!
There is a problem somewhere in the active set solver. I think that
the logic around adding and removing constraints is not correct,
which can lead to all sorts of problems.

The code is currently filled with lots of debugging print-outs and
in need of some careful work on getting the active-set part of the
algorithm correct. Best to implement directly from a good paper on
active set methods.

Pausing development on this branch for now, investigating pulling in
an existing QP solver or using the specialized solver that is in the
SNS-IK paper.